♻️ refactor: change type and update import order
