--- src/ProjectDashboard/useProjectQuery.tsx
@@ -1,17 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-
-import { useAuth } from "../@shared/useAuth";
-import APIClient from "../@utils/api";
-
-export function useProjectQuery(projectName: string) {
-  const { user } = useAuth();
-  const api = new APIClient()
-    .setUserId(user?.id)
-    .setAccessToken(user?.accessToken)
-    .setGithubAccessToken(user?.githubAccessToken);
-
-  return useQuery({
-    queryKey: ["project", projectName],
-    queryFn: () => api.getProject(projectName),
-  });
-}
